feat: #2661 moved all fake data into seperate files
