+++ apps/api/src/app/approval-policy/default-approval-policies.ts
@@ -1,0 +1,11 @@
+export const DEFAULT_APPROVAL_POLICIES: string[] = [
+	'Trade Policy',
+	'Union Budget',
+	'Definition, Licensing Policies and Registration',
+	'State Government Industrial Policies',
+	'Reservation Policy',
+	'National Policies',
+	'International Policies',
+	'Time Off',
+	'Equipment Sharing'
+];
