+++ apps/api/src/app/candidate-education/default-candidate-educations.ts
@@ -1,0 +1,10 @@
+import { ICandidateEducation } from '@gauzy/models';
+
+export const DEFAULT_CANDIDATE_EDUCATIONS: ICandidateEducation[] = [
+	{
+		schoolName: 'MIT',
+		degree: 'Master',
+		completionDate: new Date(2017, 4, 4),
+		field: 'Computer Science'
+	}
+];
