+++ apps/api/src/app/candidate-feedbacks/default-candidate-feedbacks.ts
@@ -1,0 +1,12 @@
+import { ICandidateFeedback } from '@gauzy/models';
+
+export const DEFAULT_CANDIDATE_FEEDBACKS: ICandidateFeedback[] = [
+	{
+		description: 'Feedback 1',
+		rating: 4
+	},
+	{
+		description: 'Feedback 2',
+		rating: 3
+	}
+];
