+++ apps/api/src/app/candidate-skill/default-candidate-skills.ts
@@ -1,0 +1,10 @@
+import { ISkill } from '@gauzy/models';
+
+export const DEFAULT_CANDIDATE_SKILLS: ISkill[] = [
+	{
+		name: 'Fullstack Developer'
+	},
+	{
+		name: 'MEAN Stack Developer'
+	}
+];
