+++ apps/api/src/app/candidate-source/default-candidate-source.ts
@@ -1,0 +1,19 @@
+import { ICandidateSource } from '@gauzy/models';
+
+export const DEFAULT_CANDIDATE_SOURCES: ICandidateSource[] = [
+	{
+		name: 'LinkedIn'
+	},
+	{
+		name: 'Indeed'
+	},
+	{
+		name: 'Idealist'
+	},
+	{
+		name: 'Dice'
+	},
+	{
+		name: 'Monster'
+	}
+];
