+++ apps/api/src/app/employee-appointment/default-employee-appointment.ts
@@ -1,0 +1,6 @@
+export const AGENDAS: string[] = [
+	'Meeting',
+	'Knowledge Transfer',
+	'Query Solution',
+	'Sprint Planning'
+];
