+++ apps/api/src/app/employee-award/default-employee-award.ts
@@ -1,0 +1,5 @@
+export const DEFAULT_EMPLOYEE_AWARDS = [
+	{ name: 'Microsoft Most Valuable Professional (MVP)', year: '2015' },
+	{ name: 'Google Developer Expert (GDE)', year: '2019' },
+	{ name: 'Hackatoon Winner', year: '2018' }
+];
