+++ apps/api/src/app/organization-awards/default-organization-awards.ts
@@ -1,0 +1,6 @@
+export const DEFAULT_ORGANIZATION_AWARDS = {
+	'Top Software Development Company': 2015,
+	'Upwork Top Rated Development Company 2019': 2019,
+	'Upwork Top Rated Development Company 2018': 2018,
+	'Upwork Top Rated Development Company 2017': 2017
+};
