+++ apps/api/src/app/organization-department/default-organization-department.ts
@@ -1,0 +1,10 @@
+export const DEFAULT_ORGANIZATION_DEPARTMENTS: string[] = [
+	'Designers',
+	'QA',
+	'Engineering',
+	'Management',
+	'Sales',
+	'Marketing',
+	'Frontend Developers',
+	'Backend Developers'
+];
