+++ apps/api/src/app/organization-languages/default-organization-languages.ts
@@ -1,0 +1,7 @@
+export const DEFAULT_ORGANIZATION_LANGUAGES = {
+	Bulgarian: 'Native or Bilingual',
+	Russian: 'Native or Bilingual',
+	Ukrainian: 'Native or Bilingual',
+	English: 'Fluent',
+	Hebrew: 'Conversational'
+};
