+++ apps/api/src/app/organization-positions/default-organization-positions.ts
@@ -1,0 +1,14 @@
+export const DEFAULT_ORGANIZATION_POSITIONS: string[] = [
+	'Web Software Developer',
+	'CEO',
+	'Team Lead',
+	'Staff Engineer',
+	'Junior Software Developer',
+	'Senior Software Developer',
+	'CTO',
+	'COO',
+	'Mobile Application Developer',
+	'Associate Engineer',
+	'Project Manager',
+	'Tester'
+];
