+++ apps/api/src/app/organization-positions/organization-position.seed.ts
@@ -2,21 +2,7 @@ import { Connection } from 'typeorm';
 import { OrganizationPositions } from './organization-positions.entity';
 import { Tenant } from '../tenant/tenant.entity';
 import { Organization } from '../organization/organization.entity';
-
-const organizationPositionArray = [
-	'Web Software Developer',
-	'CEO',
-	'Team Lead',
-	'Staff Engineer',
-	'Junior Software Developer',
-	'Senior Software Developer',
-	'CTO',
-	'COO',
-	'Mobile Application Developer',
-	'Associate Engineer',
-	'Project Manager',
-	'Tester'
-];
+import { DEFAULT_ORGANIZATION_POSITIONS } from './default-organization-positions';
 
 export const seedDefaultOrganizationPosition = async (
 	connection: Connection,
@@ -25,8 +11,7 @@ export const seedDefaultOrganizationPosition = async (
 ): Promise<void> => {
 	let positions: OrganizationPositions[] = [];
 
-	// organizations.forEach(({ id: organizationId }) => {
-	const organizationPositions: OrganizationPositions[] = organizationPositionArray.map(
+	const organizationPositions: OrganizationPositions[] = DEFAULT_ORGANIZATION_POSITIONS.map(
 		(name) => {
 			const employmentPosition = new OrganizationPositions();
 			employmentPosition.name = name;
@@ -36,7 +21,6 @@ export const seedDefaultOrganizationPosition = async (
 		}
 	);
 	positions = [...positions, ...organizationPositions];
-	// });
 	await insertEmploymentPosition(connection, positions);
 };
 
@@ -50,7 +34,7 @@ export const seedRandomOrganizationPosition = async (
 	for (const tenant of tenants) {
 		const organizations = tenantOrganizationsMap.get(tenant);
 		organizations.forEach(({ id: organizationId }) => {
-			const organizationPositions: OrganizationPositions[] = organizationPositionArray.map(
+			const organizationPositions: OrganizationPositions[] = DEFAULT_ORGANIZATION_POSITIONS.map(
 				(name) => {
 					const employmentPosition = new OrganizationPositions();
 					employmentPosition.name = name;
