+++ apps/api/src/app/organization-projects/default-organization-projects.ts
@@ -1,0 +1,6 @@
+export const DEFAULT_ORGANIZATION_PROJECTS: string[] = [
+	'Gauzy Platform (Open-Source)',
+	'Gauzy Web Site',
+	'Gauzy Platform SaaS',
+	'Gauzy Platform DevOps'
+];
