+++ apps/api/src/app/organization_employee-level/default-organization-employee-levels.ts
@@ -1,0 +1,16 @@
+import { IEmployeeLevelInput } from '@gauzy/models';
+
+export const DEFAULT_EMPLOYEE_LEVELS: IEmployeeLevelInput[] = [
+	{
+		level: 'Level A'
+	},
+	{
+		level: 'Level B'
+	},
+	{
+		level: 'Level C'
+	},
+	{
+		level: 'Level D'
+	}
+];
