+++ apps/api/src/app/skills/default-skills.ts
@@ -1,0 +1,8 @@
+export const DEFAULT_SKILLS: string[] = [
+	'Node.js',
+	'Angular',
+	'React',
+	'NestJS',
+	'NextJS',
+	'PostgreSQL'
+];
