+++ apps/api/src/app/skills/skill.seed.ts
@@ -1,5 +1,6 @@
 import { IOrganization, ITenant } from '@gauzy/models';
 import { Connection } from 'typeorm';
+import { DEFAULT_SKILLS } from './default-skills';
 import { Skill } from './skill.entity';
 
 export const createDefaultSkills = async (
@@ -8,16 +9,7 @@ export const createDefaultSkills = async (
 	organization: IOrganization
 ): Promise<Skill[]> => {
 	const skills: Skill[] = [];
-	const skillNames = [
-		'Node.js',
-		'Angular',
-		'React',
-		'NestJS',
-		'NextJS',
-		'PostgreSQL'
-	];
-
-	for (const name of skillNames) {
+	for (const name of DEFAULT_SKILLS) {
 		const skill = new Skill();
 		skill.name = name;
 		skill.tenant = tenant;
