+++ apps/api/src/app/tags/default-tags.ts
@@ -1,0 +1,34 @@
+export const DEFAULT_GLOBAL_TAGS: string[] = [
+	'VIP',
+	'Urgent',
+	'Crazy',
+	'Broken',
+	'TODO',
+	'In Process',
+	'Verified',
+	'Third Party API',
+	'Killer',
+	'Idiot',
+	'Super',
+	'WIP'
+];
+
+export const DEFAULT_ORGANIZATION_TAGS: string[] = [
+	'Program',
+	'Web',
+	'Mobile',
+	'Frontend',
+	'Backend',
+	'Database',
+	'Authentication',
+	'Security',
+	'Dashboard',
+	'API',
+	'Design',
+	'Testing',
+	'Local',
+	'QC',
+	'Production',
+	'Crap',
+	'WTF'
+];
