+++ apps/api/src/app/time-off-policy/default-time-off-policy.ts
@@ -1,0 +1,12 @@
+export const DEFAULT_TIMEOFF_POLICIES: string[] = [
+	'Policy 1',
+	'Policy 2',
+	'Policy 3',
+	'Policy 4',
+	'Policy 5',
+	'Policy 6',
+	'Policy 7',
+	'Policy 8',
+	'Policy 9',
+	'Policy 10'
+];
